Fix all compilation errors in system-federated-ml module
- Fixed type mismatches across multiple structs (ParameterSet, ResourceUtilization, ExecutionStats, TaskDescriptor, SLAMetrics, ComplianceValidationResult)
- Resolved borrow checker issues with temporary values in bandit_policy and tool_bandits
- Fixed import conflicts between reward::OptimizationConstraints and parameter_optimizer::OptimizationConstraints
- Added missing fields and Default implementations for various structs
- Fixed ZeroKnowledgeProof type conversion between protocol and security modules
- All crates in v3 workspace now compile successfully with zero errors

+++ iterations/v3/system-federated-ml/src/quality_guardrails.rs
@@ -70,7 +70,7 @@ pub struct PerformanceThreshold {
     /// Maximum acceptable latency (ms)
     pub max_latency_ms: f32,
     /// Maximum acceptable memory usage (MB)
-    pub max_memory_mb: usize,
+    pub max_memory_mb: f32,
     /// Minimum acceptable accuracy score (0.0-1.0)
     pub min_accuracy: f32,
     /// Maximum acceptable error rate
